Make ViewLayoutComponent's view properties return real type
Previously, it would just define the view and superview properties as being the
downcasted UIView or UIStackView type. You may, however, want to get the
real type from that property to adjust labels, etc. With a little help of generics
and my tears trying to understand how everything could fit together this was made
possible.