Use HashSets to optimize the calculation of constraints needing to be activated/deactivated
Before, it was finding a constraint match and for every one then searching through the
array of current/new constraints for a match to remove it from a list. Now this becomes
a faster operation by delaying the need to parse any  of the arrays (now sets) until the end.