Speed up removal of unneeded subviews
By using a HashSet, as it goes through the current layout's subviews,
when it asks if the new layout's subviews contains a view, that
contains operation will be a O(1) operation instead of O(n).
I have profiled this to make that process ~40% faster when running
the layout update a second time.