Improve first update performance by just activating all new constraints
If there are no current constraints activated, we don't need
to do any diffing of constraints, we can just activate the
new ones. The performance tests don't back this, but only
because the first run of a measure block seems to take
an abnormally long amount of time for no apparent reason
and makes the overall averages worthless.