Remove fast description from README
This isn't some sort of admittance the library is slow. I just don't
want to give the impression that this library will make your code faster
in any way. It will surely be faster to activate/deactivate your
constraints explicitly, but this library aims to be "fast-enough" that
the difference is negligible.