Mark all views held onto by framework as unowned
Otherwise custom UIViews that had their own ViewLayout object would
create a circular reference. It doesn't make sense for this object to
exist when the view it holds onto doesn't, so it will crash in that
scenario. If people end up having trouble debugging why things are
crashing i'll switch to using preconditions as well to give better
error messages.