Clean up example project
there was a lot of white space in the layout code that I
think was unnecessary and makes it harder to read. I also separated
out view configuration from the rest of the layout code. Even
though you can totally do it there, I want to make clear the framework
only really handles diffing of layout, not of any view properties.