Remove view and superview from being passed into component closures
This information has always been available on the component itself.
This was done to make it easier to reference these when writing
constraints. However, this made it annoying to write, because you had
to keep typing variable names for these closure arguments. There are
other patterns to make referencing these easier.